Factor out W.balanceTransaction running in `MonadRandom m`
Which can be more easily property tested.
