+++ lib/core/src/Cardano/Wallet/Api/Server.hs
@@ -356,8 +356,7 @@ import Cardano.Wallet.Primitive.AddressDiscovery.Shared
     , validateScriptTemplates
     )
 import Cardano.Wallet.Primitive.CoinSelection
-    ( SelectionCollateralRequirement (..)
-    , SelectionError (..)
+    ( SelectionError (..)
     , SelectionOf (..)
     , SelectionOutputInvalidError (..)
     , SelectionOutputSizeExceedsLimitError (..)
@@ -365,10 +364,7 @@ import Cardano.Wallet.Primitive.CoinSelection
     , selectionDelta
     )
 import Cardano.Wallet.Primitive.CoinSelection.Balance
-    ( SelectionSkeleton (..)
-    , UnableToConstructChangeError (..)
-    , balanceMissing
-    )
+    ( UnableToConstructChangeError (..), balanceMissing )
 import Cardano.Wallet.Primitive.Delegation.UTxO
     ( stakeKeyCoinDistr )
 import Cardano.Wallet.Primitive.Migration
@@ -403,7 +399,6 @@ import Cardano.Wallet.Primitive.SyncProgress
 import Cardano.Wallet.Primitive.Types
     ( Block
     , BlockHeader (..)
-    , FeePolicy (..)
     , NetworkParameters (..)
     , PassphraseScheme (..)
     , PoolId
@@ -418,11 +413,11 @@ import Cardano.Wallet.Primitive.Types
 import Cardano.Wallet.Primitive.Types.Address
     ( Address (..), AddressState (..) )
 import Cardano.Wallet.Primitive.Types.Coin
-    ( Coin (..), coinQuantity, subtractCoin, sumCoins )
+    ( Coin (..), coinQuantity )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..) )
 import Cardano.Wallet.Primitive.Types.Redeemer
-    ( Redeemer (..), redeemerData )
+    ( Redeemer (..) )
 import Cardano.Wallet.Primitive.Types.TokenBundle
     ( Flat (..), TokenBundle (..) )
 import Cardano.Wallet.Primitive.Types.TokenMap
@@ -430,15 +425,13 @@ import Cardano.Wallet.Primitive.Types.TokenMap
 import Cardano.Wallet.Primitive.Types.TokenPolicy
     ( TokenName (..), TokenPolicyId (..), nullTokenName )
 import Cardano.Wallet.Primitive.Types.Tx
-    ( SealedTx
-    , TransactionInfo
+    ( TransactionInfo
     , Tx (..)
     , TxChange (..)
     , TxIn (..)
     , TxOut (..)
     , TxStatus (..)
     , UnsignedTx (..)
-    , txOutAddCoin
     , txOutCoin
     )
 import Cardano.Wallet.Registry
@@ -456,16 +449,11 @@ import Cardano.Wallet.Transaction
     , ErrSignTx (..)
     , TransactionCtx (..)
     , TransactionLayer (..)
-    , TxUpdate (..)
     , Withdrawal (..)
     , defaultTransactionCtx
     )
 import Cardano.Wallet.Unsafe
     ( unsafeRunExceptT )
-import Cardano.Wallet.Util
-    ( mapFirst )
-import Control.Applicative
-    ( (<|>) )
 import Control.Arrow
     ( second )
 import Control.DeepSeq
@@ -593,7 +581,6 @@ import UnliftIO.Concurrent
 import UnliftIO.Exception
     ( IOException, bracket, throwIO, tryAnyDeep, tryJust )
 
-import qualified Cardano.Api.Shelley as Cardano
 import qualified Cardano.Wallet as W
 import qualified Cardano.Wallet.Api.Types as Api
 import qualified Cardano.Wallet.Network as NW
@@ -602,7 +589,6 @@ import qualified Cardano.Wallet.Primitive.AddressDerivation.Icarus as Icarus
 import qualified Cardano.Wallet.Primitive.CoinSelection.Balance as Balance
 import qualified Cardano.Wallet.Primitive.CoinSelection.Collateral as Collateral
 import qualified Cardano.Wallet.Primitive.Types as W
-import qualified Cardano.Wallet.Primitive.Types.Coin as Coin
 import qualified Cardano.Wallet.Primitive.Types.TokenBundle as TokenBundle
 import qualified Cardano.Wallet.Primitive.Types.TokenMap as TokenMap
 import qualified Cardano.Wallet.Primitive.Types.Tx as W
@@ -615,7 +601,6 @@ import qualified Data.Aeson as Aeson
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as BL
 import qualified Data.Foldable as F
-import qualified Data.List as L
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
 import qualified Data.Set as Set
@@ -2243,188 +2228,23 @@ balanceTransaction ctx genChange (ApiT wid) body = do
     pp <- liftIO $ NW.currentProtocolParameters nl
     -- TODO: This throws when still in the Byron era.
     nodePParams <- fromJust <$> liftIO (NW.currentNodeProtocolParameters nl)
-
-    ti <- liftIO $ snapshot $ timeInterpreter (ctx ^. networkLayer)
-
-    let (outputs, txWithdrawal, txMetadata, txAssetsToMint, txAssetsToBurn) = extractFromTx partialTx
-
-    (delta, extraInputs, extraCollateral, extraOutputs) <-
-      withWorkerCtx ctx wid liftE liftE $ \wrk -> do
-        (internalUtxoAvailable, wallet, pendingTxs) <-
-            liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
-
-        let externalSelectedUtxo = UTxOIndex.fromSequence
-                ((\(a,b,_)-> (a,b)) <$> externalInputs)
-
-        let utxoAvailableForInputs = UTxOSelection.fromIndexPair
-                (internalUtxoAvailable, externalSelectedUtxo)
-
-        let utxoAvailableForCollateral =
-                UTxOIndex.toUTxO internalUtxoAvailable
-
-        -- NOTE: It is not possible to know the script execution cost in
-        -- advance because it actually depends on the final transaction. Inputs
-        -- selected as part of the fee balancing might have an influence on the
-        -- execution cost.
-        -- However, they are bounded so it is possible to balance the
-        -- transaction considering only the maximum cost, and only after, try to
-        -- adjust the change and ExUnits of each redeemer to something more
-        -- sensible than the max execution cost.
-        let txPlutusScriptExecutionCost = maxScriptExecutionCost tl pp redeemers
-        let txContext = defaultTransactionCtx
-                { txPlutusScriptExecutionCost
-                , txMetadata
-                , txWithdrawal
-                , txAssetsToMint
-                , txAssetsToBurn
-                , txCollateralRequirement =
-                    if txPlutusScriptExecutionCost > Coin 0 then
-                        SelectionCollateralRequired
-                    else
-                        SelectionCollateralNotRequired
-                } & padFeeEstimation partialTx pp nodePParams
-
-        -- FIXME: The coin selection and reported fees will likely be wrong in
-        -- the presence of certificates (and deposits / refunds). An immediate
-        -- "fix" is to return a proper error from the handler when any key or
-        -- pool registration (resp. deregistration) certificate is found in the
-        -- transaction. A long-term fix is to handle this case properly during
-        -- balancing.
-        let transform s sel =
-                let (sel', _) = W.assignChangeAddresses genChange sel s
-                    inputs = F.toList (sel' ^. #inputs)
-                 in ( selectionDelta txOutCoin sel'
-                    , inputs
-                    , fst <$> (sel' ^. #collateral)
-                    , sel' ^. #change
-                    )
-        liftHandler $ W.selectAssets @_ @_ @s @k wrk pp W.SelectAssetsParams
-            { outputs
-            , pendingTxs
-            , txContext
-            , utxoAvailableForInputs
-            , utxoAvailableForCollateral
-            , wallet
-            }
-            transform
-
-    -- NOTE:
-    -- Once the coin-selection is done, we need to
-    --
-    -- (a) Add selected inputs, collateral and change outputs to the transaction
-    -- (b) Assign correct execution units to every redeemers
-    -- (c) Correctly reference redeemed entities with redeemer pointers
-    -- (d) Adjust fees and change output(s) to the new fees.
-    --
-    -- There's a strong assumption that modifying the fee value AND increasing
-    -- the coin value of change outputs does not modify transaction fees; or
-    -- more exactly, does not modify the execution units of scripts. This is in
-    -- principle a fair assumption because scripts validators ought to be
-    -- unaware of change outputs. If their execution cost increases when change
-    -- output increases, then it becomes impossible to guarantee that the fee
-    -- balancing will ever converge towards a fixed point. A script validator
-    -- doing such thing is considered bonkers and this is not a behavior we
-    -- ought to support.
-
-    candidateTx <- assembleTransaction nodePParams ti $ TxUpdate
-        { extraInputs
-        , extraCollateral
-        , extraOutputs
-        , newFee = const delta
-        }
-    let candidateMinFee = fromMaybe (Coin 0) $
-            evaluateMinimumFee tl nodePParams candidateTx
-
-    let surplus = delta `Coin.distance` candidateMinFee
-    finalTx <- assembleTransaction nodePParams ti $ TxUpdate
-        { extraInputs
-        , extraCollateral
-        , extraOutputs = mapFirst (txOutAddCoin surplus) extraOutputs
-        , newFee = const candidateMinFee
-        }
-
-    pure $ ApiSerialisedTransaction
-        { transaction = ApiT finalTx
-        }
+    let partialTx = W.PartialTx
+            (getApiT $ body ^. #transaction)
+            (fromExternalInput <$> body ^. #inputs)
+            (fromApiRedeemer <$> body ^. #redeemers)
+    withWorkerCtx ctx wid liftE liftE $ \wrk -> do
+        wallet <- liftHandler $ W.readWalletUTxOIndex @_ @s @k wrk wid
+        ti <- liftIO $ snapshot $ timeInterpreter $ ctx ^. networkLayer
+        transaction <- liftHandler $ ApiT <$> W.balanceTransaction @IO @s @k
+            wrk
+            genChange
+            (pp, nodePParams)
+            ti
+            wallet
+            partialTx
+        return $ ApiSerialisedTransaction { transaction }
   where
     nl = ctx ^. networkLayer
-    tl = ctx ^. W.transactionLayer @k
-
-    partialTx :: SealedTx
-    partialTx = body ^. #transaction . #getApiT
-
-    redeemers :: [Redeemer]
-    redeemers = fromApiRedeemer <$> body ^. #redeemers
-
-    externalInputs :: [(TxIn, TxOut, Maybe (Hash "Datum"))]
-    externalInputs = fromExternalInput <$> body ^. #inputs
-
-    assembleTransaction
-        :: Cardano.ProtocolParameters
-        -> TimeInterpreter (Either PastHorizonException)
-        -> TxUpdate
-        -> Handler SealedTx
-    assembleTransaction nodePParams ti update = do
-        tx' <- asHandler $ updateTx tl partialTx update
-        liftHandler $ ExceptT $ pure $ assignScriptRedeemers
-            tl nodePParams ti resolveInput redeemers tx'
-      where
-        resolveInput :: TxIn -> Maybe (TxOut, Maybe (Hash "Datum"))
-        resolveInput i =
-            (\(_,o,d) -> (o,d)) <$> L.find (\(i',_,_) -> i == i') externalInputs
-            <|>
-            (\(_,o) -> (o, Nothing)) <$> L.find (\(i',_) -> i == i') (extraInputs update)
-
-    extractFromTx tx =
-        let (Tx _id _fee _coll _inps outs wdrlMap meta _vldt, toMint, toBurn) = decodeTx tl tx
-            -- TODO: Find a better abstraction that can cover this case.
-            wdrl = WithdrawalSelf
-                (error $ "WithdrawalSelf: reward-account should never been use "
-                      <> "when balancing transactions but it was!"
-                )
-                (error $ "WithdrawalSelf: derivation path should never been use "
-                      <> "when balancing transactions but it was!"
-                )
-                (sumCoins wdrlMap)
-         in (outs, wdrl, meta, toMint, toBurn)
-
-    -- | Wallet coin selection is unaware of many kinds of transaction content
-    -- (e.g. datums, redeemers), which could be included in the input to
-    -- `balanceTransaction`. As a workaround we add some padding using
-    -- `evaluateMinimumFee`.
-    --
-    -- TODO: This logic needs to be consistent with how we call `selectAssets`,
-    -- so it would be good to join them into some single helper.
-    padFeeEstimation
-        :: W.SealedTx
-        -> W.ProtocolParameters
-        -> Cardano.ProtocolParameters
-        -> TransactionCtx
-        -> TransactionCtx
-    padFeeEstimation sealedTx pp pp' txCtx =
-        let
-            (walletTx, _, _) = decodeTx tl sealedTx
-            worseEstimate = calcMinimumCost tl pp txCtx skeleton
-            skeleton = SelectionSkeleton
-                { skeletonInputCount = length (view #resolvedInputs walletTx)
-                , skeletonOutputs = view #outputs walletTx
-                , skeletonChange = mempty
-                }
-            LinearFee _ (Quantity b) = pp ^. #txParameters . #getFeePolicy
-            -- NOTE: Coping with the later additions of script integrity hash and
-            -- redeemers ex units increased from 0 to their actual values.
-            extraMargin = Coin $ ceiling $ (*) b $ fromIntegral
-                $ sizeOfScriptIntegrityHash
-                + sum ((+sizeOfRedeemerCommon) . BS.length . redeemerData <$> redeemers)
-              where
-                sizeOfScriptIntegrityHash = 35
-                sizeOfRedeemerCommon = 17
-
-            txFeePadding = (<> extraMargin) $ fromMaybe (Coin 0) $ do
-                betterEstimate <- evaluateMinimumFee tl pp' sealedTx
-                betterEstimate `subtractCoin` worseEstimate
-        in
-            txCtx { txFeePadding }
 
 joinStakePool
     :: forall ctx s n k.
@@ -3665,11 +3485,6 @@ liftHandler action = Handler (withExceptT toServerError action)
 liftE :: IsServerError e => e -> Handler a
 liftE = liftHandler . throwE
 
-asHandler :: IsServerError e => Either e a -> Handler a
-asHandler = \case
-    Left e  -> liftE e
-    Right a -> return a
-
 apiError :: ServerError -> ApiErrorCode -> Text -> ServerError
 apiError err code message = err
     { errBody = Aeson.encode $ Aeson.object
@@ -3880,6 +3695,8 @@ instance IsServerError ErrBalanceTx where
                 , "the transaction body is modified. "
                 , "Please sign the transaction after it is balanced instead."
                 ]
+        ErrBalanceTxSelectAssets err -> toServerError err
+        ErrBalanceTxAssignRedeemers err -> toServerError err
         ErrBalanceTxNotImplemented ->
             apiError err501 NotImplemented
                 "This feature is not yet implemented."
